+++ templates/landing.php
@@ -12,7 +12,7 @@ add_filter( '_mai_header_before', '__return_false' );
 add_filter( '_mai_header_left', '__return_false' );
 add_filter( '_mai_header_right', '__return_false' );
 add_filter( '_mai_header_after', '__return_false' );
-add_filter( '_mai_mobile_menu', '__return_false' );
+add_filter( '_mai_mobile_menu', '__return_false' ); // Mobile menu kill switch. In header.php also.
 
 // Remove navigation.
 remove_action( 'genesis_after_header', 'genesis_do_nav' );